Reorganize Seth repo
The sawtooth-seth repo was separated from sawtooth-core by deleting
out all non-Seth related documents and files. This has led to a
disorganized and confusing directory structure. This commit takes steps
towards reorganizing the repo.

Warning: Protogen and build scripts may not be functional in this commit
due to changed paths.

Signed-off-by: Darian Plumb <dplumb@bitwise.io>