Mount contracts directory in cli image
This links the sawtooth-seth/contracts directory to
/project/sawtooth-seth/contracts in the seth-cli container, so any
files in this directory will be visible and editable on both the host
and inside the container.

The purpose is to allow a user to create and edit contracts on the host
and compile them while inside the container using the installed
compilation and abi generation tools.

Example contracts are included in the repo, but other user-created
contracts will be ignored by git.

Signed-off-by: Darian Plumb <dplumb@bitwise.io>