Share key dir between cli and rpc containers
The Seth RPC container relies on keys that were generated by the CLI.
Previously, we were generating those keys inside of the RPC container
with a CLI call. This PR moves to sharing a key directory between the
CLI and RPC containers, with the CLI container generating the keys in
that directory.

Signed-off-by: Kenneth Koski <knkski@bitwise.io>