Wait to run tests until rest-api is up
The rest-api service is a bit slow to start up, and if it has not
started accepting connections by the time the "go test foo.go" command
runs, tests will fail. This PR adds a wrapper script for the tests that
waits for the rest api to be up and responsive, then runs them as
normal.

Signed-off-by: Kenneth Koski <knkski@bitwise.io>